Finished the Required Views
The 6 views made:
____________________
- home
- dog_product_detail
-purchase_dog_product
-purchase_detail
- new_dog_tag
- dog_tag_list

Quick Overview:
_________________
- home :
This is more or less the landing page and aids in showing all the dog products available to the user.

- dog_product_detail:
This aids the page understand which dog product the user decided to click on to read more on. It does this with the help of the `dog_product_id`.

-purchase_dog_product:
This aids the page have the functionality of when you click `buy` on the product it changes the quantity and helps build up the view counter part `purchase_detail`. It also has a small amount of fault tolerence by not allowing customers buy a product that is out of stock.

-purchase_detail:
This aids the page display the item the customer purchased along as the rest of the purchase details.

-new_dog_tag:
This view has 2 uses. If the user wants to make a `new_dog_tag` then it hands them the page with the form. Otherwise It will take the users input and add it the `dog_tag_list`.

-dog_tag_list:
This aids the page with displaying all the dog tags the user has made.