Update title to something more apporiate
I decided to make this change because I wanted to create a more accurate README based on the work I did rather than the steps I took to get there. This is still very valid information so that is why I am changing it to another file and adding a new README later.